* checkbashisms:   + Apply a small fix to part of the "quoted block detection"
This reverts part of r1441. The bug it was originally introduced to fix was actually a side-effect of 
a bug elsewhere and the cure in this case was actually worse than the disease. (For reference, the 
real problem lies in the comment removal code and how it deals with quoted strings)