uscan: properly cope with colons and minus signs in upstream versions
uscan uses dpkg --compare-versions and Dpkg::Version to compare upstream
versions.  But this command takes debian versions, which means it produces
wrong results if the upstream version contains colons or minus signs.

For example    3-1-1-3   >   3-1-4   (as 'upstream' is 3-1-1   vs. 3-1)
but            3-1-1-3-1 <   3-1-4-1 (as 'upstream' is 3-1-1-3 vs. 3-1-4)

To work around this compare the upstream versions with a epoch before
and a revision afterwards, so colons and minus signs are treated as part
of the upstream version.

Calls to dpkg --compare-versions in uscan get this treatment directly,
for the Dpkg::Version calls fix Devscripts::Versort:

Add Devscripts::Versort::upstream_versort to sort versions ad upstream
versions. (Devscripts::Versort::versort was documented to take upstream
versions, but some other scripts seem to call it with Debian versions,
so keep that behaviour and add a new one).

Signed-off-by: Bernhard R. Link <brlink@debian.org>
Signed-off-by: James McCoy <jamessan@debian.org>