rc-alert: Unset wget's continue option
If a user has "continue = on" in their wgetrc, then when wget determines the
remote file needs to be downloaded, it will only download enough to make up
the file size  difference.  This results in an invalid cached bug list.
Disabling continue ensures that when a download happens, the entire file is
downloaded.

Closes: #677229
Signed-off-by: James McCoy <jamessan@debian.org>