new script origtargz.pl
origtargz downloads the orig tarball of a Debian package, and optionally
unpacks it into the current directory. The version number to be used is
determined from debian/changelog. The main use for origtargz is with
debian-dir-only repository checkouts. It should be invoked from the top
level directory of an unpacked Debian source package.