checkbashisms: do not mistake bash arrays as functions
"foo=()" would trigger the function names check even though it is
an empty array being assigned to variable "foo".

Forbid the function names regex from matching if the last character in
what appears to be the function name is the equality sign. In order to
still detect a function called "=" add an extra check that requires the
sign to be proceeded by a blank or be located at the start of the line.

Signed-off-by: Benjamin Drung <bdrung@debian.org>