dscverify: Use "gpg --status-fd" to get more details about validity
Simply running "gpg < file" doesn't ensure the content is properly
signed.  Even when it does, we may not be using the signed content.

Using "gpg --status-fd 1 < file" solves both of these issues.  Even
though it still won't error out with an unsigned file, we'll be able to
detect that the content wasn't signed by the lack of a VALIDSIG status.
Also, the command will emit the signed content between PLAINTEXT status
and any subsequent status lines.

Closes: #695914
Signed-off-by: James McCoy <jamessan@debian.org>