uscan: Define local replacements for die/warn.
Setting $SIG{__DIE__}/$SIG{__WARN__} in a non-localized manner has
effects on other modules that we may call.  This action at a distance
can lead to broken behavior, such as exiting uscan when an eval'd die
occurs in some support module.

Define and use uscan_die/uscan_warn as replacements for die/warn which
will act like normal if --dehs isn't given and provide diagnostics
through the dehs output when it is.

Closes: #669942
Signed-off-by: James McCoy <jamessan@debian.org>