Fix various (potential) test failures.
Some devscripts commands can be configured by a config file.  If that is
present in the test environment, it could cause a false negative test
failure, as I noticed with the new uscan test.  Use --no-conf to ignore
the config files.

The online uscan test was also failing due to a change in the format of
the version numbers used by the upstream authors, which is fixed by
broadening the allowable version characters.

Signed-off-by: James McCoy <jamessan@debian.org>