rc-alert: Parse a package div at a time
The HTML parsing is currently very reliant on the specific format that
bugs.d.o sends.  If this is manipulated in any way (e.g. removal of
newlines) then the parsing breaks.

To remedy that, slurp the entire page instead of working a line at a
time.  Then it is trivial to extract the contents of each <div
class="package"> stanza and parse out the relevant information.

Closes: #729779
Signed-off-by: James McCoy <jamessan@debian.org>