uscan: Implement Files-Excluded via "tar --delete"
The code is cleaner, less expensive file operations (although we still
decompress and compress via a temporary file) and the compression method
is unchanged.

If the file is not a tarball originally (e.g. a zip file), this does not
work any more. But shouldn't in that case --repack be explicitly asked
for by the user?