Allow more flexibility for specialised build chroots
Debian now has backports archives as an Official Thing, and sometimes
packages which are to be built for them depend on other packages that
are already in them.  But they aren't fully fledged suites in their
own right, so setting up a cowbuilder chroot for them isn't the same
as for an ordinary release.  Rather it has more in common with having
a local repository of extra packages for a given suite.

Since cowpoke has until now required 'dist' to be a real suite that
debootstrap recognised, that made it a bit awkward to have both a
pristine chroot for normal package builds intended for upload to one
of the official suites, and a 'special' chroot with extra packages
or some other oddball configuration, on the same build host if the
needed suites overlapped.  So we add the concept of a BASE_DIST now
and decouple the 'dist' name from the official suites, which means
you are able to define as many chroots as you please, with whatever
configuration you please, without having to destroy and remake them
when you need a pristine one again for some suite.

We also add options to allow arbitrary cowbuilder/pbuilder options
to be passed through during the --create, --update, and --build
phases of its operation.  Which makes the otherwise tedious task of
adding additional repositories or packages, or build options, just
as trivial as any normal use of cowpoke is.

Allow SIGN_KEYID and UPLOAD_QUEUE to be per arch/dist too

Since now that it's trivial to have multiple chroot flavours, and we
aren't all that far from people all having their own Probably Putrid
archives to upload to, and since that's already something many people
do for private or work (in progress) packages, it makes sense to make
that easier to not screw up.  We really don't want people uploading
bogus packages to the distro proper accidentally, or signing packages
with their distro key that aren't intended for upload to it either
just because doing something other than that is A Bit Of A Hassle.

We still don't ever offer to upload unsigned packages at all, the game
here is to make doing The Right Thing easier than not doing it. If you
really want to be stupid, you're going to have to work for that still.

Better handling of --debbuildopts

By the power of arrays vested in me, we avoid a few of the stranger ways
which this might choke on the input of an unsuspecting, but otherwise
well meaning, innocent user.  You'd probably have to try pretty hard to
hit this directly from the command line, but once we start to get called
from other layers of automation, all bets were off for sensible input at
the far end still looking as sensible once it finally got to dpkg-bp.