cowpoke: Pass --distribution to cowbuilder for --update too
Because if you pass --override-config in UPDATE_OPTS without it, then
apparently pbuilder will happily default it to sid for you, whatever
the chroot was before that.  It just ignores it otherwise though, so
it should be safe to simply do this all the time.

It shouldn't be often that somebody wants to do that, but it can be
useful for adding an extra --othermirror or similar to an existing
chroot without having to completely reinstall it.