cowpoke: Add --sign and --upload command line overrides
Now that gitpkg allows overriding cowpoke options per-invocation, it is
more useful to have these too for people building branches for backports
or local package repositories.  This is the logical extension of the
changes in 396671470242827a73dc55f71cee9714df00d5d0, which allowed these
options to be set per arch/dist in the configuration file.