explicitly dearmor, rather than --import upstream signing key
gnupg 2.1 handles --import by generating a pubring.kbx file, not a
pubring.gpg file.  So when gnupg 2.1 is installed verifying upstream
signatures failed when the upstream signing key was ascii-armored.

With this patch, we explicitly dearmor the .asc, rather than relying
on side effects that aren't true for all gpg versions.

Signed-off-by: James McCoy <jamessan@debian.org>