Call dpkg-architecture with any and all -a/-t switches
In 2.14.9/2.14.10, scripts were changed to pass on the -a/-t switches
only when both were present to fix a bug with not specifying a value for
those switches.  However, that broke cases where only one of the
switches was provided.

This commit handles all scenarios (-a & -t specified, -a specified, -t
specified, none specified) which is inline with the original behavior.

Closes: #768587
Signed-off-by: James McCoy <jamessan@debian.org>