uscan: Use URI & URI::QueryParam to parse URIs
The ordering of the query parameters in debbug's URLs isn't stable any
more (and we shouldn't have relied on that anyway).  This broke the
manual parsing of web pages when trying to cache parts of a bug.

Using URI to programatically extract the parameters simplifies the code
and avoids the issue of parameter ordering.

Closes: #786706
Signed-off-by: James McCoy <jamessan@debian.org>