licensecheck: fix digia © and license extraction (Closes: #789074)
in Digia's case, the owner contains "or", so this keyword was removed
from $copyright_disindicator_regex

I've also fixed the license extraction from Digia file. This was broken
in 2.15.4 and 2.15.3 (and possibly before, but I did not test)

The fix consists in tweaking the regex and testing alternate versions
before unique versions.