licensecheck: filter scanned files by mime type
Currently, licensecheck -r uses find to scan a directory and accepts files
based on their suffix (i.e. accepts .c .h .cxx ...)

This list of suffixes is a big regexp that must be updated regularly. Still
some files are missed like config.guess.

Maintaining this regexp is not efficient.

With this commit, licensecheck use 'file' command to decide whether
to scan a file or not. All files of mime type 'text/*' and
'application/xml' are scanned.

Note that file is already used to find the charset of each scanned file,
so there's no performance impact.