licensecheck: store only first copyright block
Some files put © info in comment and then the same info in a variable.
In this case licensecheck provides duplicated information.

Once we assume © information is specified in a single block, i.e. all
© stanza are grouped and not separated by more than 5 lines without
© information, the second block can be detected and discarded.

The number of lines between 2 blocks (5) is rather arbitrary and can be
changed if proven wrong.