uscan: new oversionmangle rule
oversionmangle:
    the further mangling of uversionmangled version to the pristine
    tarball version

test_uscan_mangle: enable test for oversionmangle

Use case:
If you wish to match the downloaded <file>-<version>.tar.gz to be the
same as the actual upstream, uversionmangle can not be used to prefix
version number such as 20010101 with prefix 0. for
<packagename>_<version>.orig.tar.gz.  (This trick enables to avoid epoch
in future)

Also for the multi upstream tarball (MUT) case, conditional suffix does
not work since all repackaged and non-repackaged packages needs to carry
the same suffix.  Also the last version may use +dfsg1, +dfsg2, or ...
so, you need to have dversionmangle to remove it for comparison with the
downloaded file.  There is a need to readd suffix for all tarballs..

similar feature origverrsionmangle was discussed as:
 https://bugs.debian.org/458777