uscan.pl: manpage update for compression etc.
https://bugs.debian.org/775526
/usr/bin/uscan: passing --compression=xz ignored

Fixed typos etc.

Remove deprecated pstream-signing-key.pgp and add comment to the source.

Under mangle mode, uscan checks and warn possible GPG keys but does not use it
automatically.  There were bug on this caused by my change to manpage and it is
corrected with this commit.

Also, the following was already addressed in the initial POD commit.
https://bugs.debian.org/573631
[uscan] manpage diverges from actual behaviour