uscan: reorganize --safe, --download, --signature
This is continuation of e3549cdbc31be498559b8bde217bfabcf4c6eb9b
    uscan: reorganize around $download etc.]

Since the difference between --no-download and --report was
ambiguous, --safe is added to clarify and stress the fact this is
guaranteed not to run any unsafe scripts.

Also, reorganized --download and --signature related logic further
and improved messaging.

 * Add USCAN_SAFE and --safe
 * Make --report alias of --safe
 * Review --download, --signature, and --safe
 * Fix --overwrite-download
 * Use --skip-signature as default for --no-download
 * process mk-origtargz not in pgpmode=next but
   in pgpmode=previous and reorganize.
 * --package should process mk-origtargz if --download

mk-origtargz currently do not run any unsafe script.  But
if Bug #748474 is addressed, this may be risky.
Thus --safe should not run mk-origtargz.