uscan: filepattern without (.*)
Fix #526450 and #803948: Please support version pattern matching in any
position of the URL

If filepattern is without (.*)  match excliding the non-capturing
(?:.*) match, we need to generate the downloaded filename with the
version from the full download URL.  This introduces the state
variable $versionless.

Since the full download URL is needed, move the filenamemangle code
after the downloadurlmangle one to use $upstream_url if needed
instead of $newversion.