dget: Match any priority for candidate version
When parsing the output of “apt-cache policy $package”, we choose the
"Candidate" version of the package to download.  Then we try to find the
matching portion of the output that lists the URLs where the package can
be retrieved.

In older versions of apt, this always listed a priority of 0, but with
apt 1.1 the priority is properly reported.  Simply changing the match
from 0 to any number allows dget to find the URIs again.

Closes: #807064
Signed-off-by: James McCoy <jamessan@debian.org>