uscan: fix prioritizing mechanism of package files
There were regressions and glitches in the prioritizing mechanism of
packages.

The new prioritizing mechanism of package files for the compression
suffix adds the suffix to the upstream version as if it is a debian
version and sorts the version with Devscripts::Versort::version instead
of Devscripts::Versort::upstream_version.

Devscripts::Versort::deb_version is not used here, since it checks
sanity of the version.

Git tag should be the upstream version.

Closes: #809604, #812417