debsnap: Fix logic when tracking if given arch had no binary
If a non-desired architecture had multiple binaries for a given version
(e.g., e2fsck-static on armhf for version 1.42-1), this would confuse
the code trying to detect when desired architectures didn't have a
binary.

Ensure that the architecture being filtered is one that we are tracking
before modifying the tracking data, and change the counting to be how
many binaries will be downloaded.  Therefore, any architecture with a 0
count is missing.

Signed-off-by: James McCoy <jamessan@debian.org>