debuild: Don't set a hook to just "cd ..; "
In 3e71ba36160b9fb5d03d755f94da14093e742a4b, code was added to ensure
debuild ran hooks from the correct directory (in source or in parent)
per the expected behavior.  However, that was also causing hooks to be
set to "cd ..; " if it was a hook that required directory adjustment but
there was no hook command set.

Ensure a hook command exists first, so we don't have the harmless, but
odd, hook command set.

Signed-off-by: James McCoy <jamessan@debian.org>