debuild: Run lintian (and hook) directly
When the running of lintian was moved to dpkg-buildpackage, that removed
the ability to ignore the return code.  While it's generally better to
avoid lintian having a non-zero return code, this was an unplanned
regression and breaks valid use cases.

Similarly, the check for the existence of lintian was restored, so
debuild doesn't fail if lintian doesn't exist.

Signed-off-by: James McCoy <jamessan@debian.org>