git-deborig: Refactor to permit archiving branches
git-archive(1) can archive both branches and tags.  Refactor the script
so that when the user manually specifies what they want to archive, it
can be a tag or a branch.

When the user does not specify what to archive, we still consider only
tags.  Branch heads move, and the script should be idempotent unless
instructed.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>