debsign: Determine related filenames from .changes/.buildinfo
Rather than trying to guess at the filenames by munging the name of the
file we were given, just inspect its contents.  This ensures we're
signing the correct files, instead of unrelated files that happen to
match the munging we're doing, as well as fixing are broken guessing for
buildinfo filenames.

Signed-off-by: James McCoy <jamessan@debian.org>