debsign: Refactor remote file handling
Restore use of $remotedir to specify where to copy files on the remote
host.  This is needed to allow glob patterns for the remote changes
file.

This also allows removing all the extra "$remote*" arguments to the
maybesign_* functions.  All that's required is the host and directory.

Finally, lift the ensure_local_copy calls out of maybesign_* functions,
since the file needs to be local in order to derive the child filenames.

Signed-off-by: James McCoy <jamessan@debian.org>