d/control: Promote sensible-utils from Recommends to Depends
Main reason is to shut up lintian's
script-needs-depends-on-sensible-utils, but it makes sense anyway as
it's a fairly important/used package that everybody should use u.U.

Previously, it was always installed as it was a dependency of an
Essential:yes package.

In particular, it's needed by nmudiff, also very important to everybody!

Signed-off-by: Mattia Rizzolo <mattia@debian.org>