debchange: unify version information
instead of having a guessed version info per type of upload (stable,
bpo, but also eventually lts, security, etc), we just have a single
variable we use.

we also merge the bpo and stable information hashes because they
mostly contain the same data. only stable had less information, but if
the info wasn't found, it was falling back to the previous entry in
the changelog, which is not better than our guess.

this also partly fixes -s uploads distro guessing.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>