debchange: twak the "automatic set targe distribution" regex so it can matches also backports
Looking at that regex now I doubt it ever matched anything before (even
before this little refactoring), as it used to check for 'bpo(\d+)\+$'
or 'deb(\d+)u$', i.e. a version ending with + or u.  Also the logic
after that feels fishy to me.   But let's keep this chunk of code around
until somebody gets the time to do a proper, real overhoul of
everything.

Signed-off-by: Mattia Rizzolo <mattia@debian.org>