scripts/bts: add tag list to manpage
I keep on having to go back to the online BTS documentation, which
doesn't work when I'm offline. That reference is also hard to find, so
I add reference to the website.

We just keep a flat list of the latest tags, without explanation, to
avoid too much duplication.

This duplicates the @gtag variable, but there is no way to interpolate
variables in POD strings. We *could* do the opposite (extract the
variable from a POD string) but that would be a mess because the list
would be at the top of the file...