uscan: redirect alioth gemwatch to gemwatch.debian.net
The Ruby team used to run a uscan redirector at
https://pkg-ruby-extras.alioth.debian.org/cgi-bin/gemwatch

It is not simply a redirector; it also converts Ruby .gem packages to
tarballs for Debian packaging, so it can't just be folded into e.g. the
qa.debian.org redirector.

This service was migrated to https://gemwatch.debian.net/

Right now the original CGI script was turned into a redirector to the
new service, and new packages produced by dh-make-ruby will already
point to it. However, we still have 900+ debian/watch files pointing to
the old URL, and the redirector will cease to work soon when alioth goes
away. The existing references will be migrated, but it is not feasible
to complete this migration process before the alioth shutdown.

This is a mitigation measure. It can and should be removed when the
migration is finished.