sadt: Add support for autopkgtests that are marked as flaky
Flaky tests are intended to pass, but are known to be unreliable,
subject to intermittent failures or otherwise unsuitable for gating CI.
To determine when a flaky test has become reliable, it is useful to
run it in realistic infrastructure for a while and gather data.

If a flaky test passes it is reported in the same way as any other
passing test, but if a flaky test fails (for any reason) it is
reported as flaky instead of as a failure.

As suggested in <https://bugs.debian.org/851558>
and implemented for autopkgtest in
<https://salsa.debian.org/ci-team/autopkgtest/merge_requests/19>.

Signed-off-by: Simon McVittie <smcv@debian.org>