sadt: Add support for autopkgtests being marked as skippable
A skippable test probes for some resource that it needs, which might
be something too complicated to express in Restrictions (for example
"/var/tmp supports extended attributes" for flatpak and ostree). If
that resource is not found, the test exits 77 (a convention borrowed
from Automake) and the test runner treats it as though it had been
skipped based on its Restrictions.

As suggested in <https://bugs.debian.org/851558>
and implemented for autopkgtest in
<https://salsa.debian.org/ci-team/autopkgtest/merge_requests/20>.

Signed-off-by: Simon McVittie <smcv@debian.org>