debsign: Fix signing when an at sign is in the path
debsign does not correctly determine the size of the .buildinfo file if
the path contains an at sign followed by a number, because the filename
is not passed literally to Perl and will be interpolated. Example:

```
$ child='/tmp/test@2'
$ touch "$child"
$ perl -i -pe 'BEGIN { '"\$file=\"$child\"; "' print $file . "\n"; }' $child
/tmp/test
```

This leads to a missing .buildinfo size in the .changes file and dput
will complain about this malformed file.

Therefore pass the filename and the checksums literally to Perl.

Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>