sadt: Ignore new restriction 'superficial'
In autopkgtest, superficial tests are the opposite of the 'flaky'
restriction: if the test fails, it is counted as a failure, but if it
succeeds, it is counted as though it had been skipped. This allows
autopkgtest to give a different exit status (8 rather than 0) if there
is no real evidence that the package works, because every test was
either skipped or superficial.

The intention is that trivial smoke tests, such as linking a
"hello, world" program to a library without exercising the
library's real functionality, can be marked as superficial.
See https://bugs.debian.org/904979 for more details. The restriction
was originally proposed as "trivial" but was renamed to "superficial"
during review.

sadt is a simpler test runner and does not distinguish between
successful and skipped tests, so we can treat superficial and
non-superficial tests identically here.

Signed-off-by: Simon McVittie <smcv@debian.org>