tests: Fix autopkgtest
Fix autopkgtest that was failing due to missing test dependency on
fakeroot.

It would be cool to specify Rules-Requires-Root:no in the test fixtures,
but test_ppackage_lifecycle compares the dpkg-buildpacakge output, and
it's a pain to do now as we'd need to change the expected output for
stretch vs buster.  Just do it once we are not interested in stretch
anymore.

Closes: #910262
Signed-off-by: Mattia Rizzolo <mattia@debian.org>