Fix the trap handling in multiple shell scripts.
Now they trap only EXIT, and not a random set of the first 15 signals.

Trapping a signal also means the trap should deal with properly exiting
the script, which wouldn't happen in any of these cases as none of the
trap called `exit`.  In all those traps what we want is to just clean up
the tempfiles while quitting the script, so just catching EXIT is fine.

Closes: #911969
Signed-off-by: Mattia Rizzolo <mattia@debian.org>