debi: Consider installing foreign 'Multi-Arch: same' packages
Multi-Arch: same packages need to be installed in lockstep.

For example, if you have a merged .changes file for dbus on amd64,
i386 and s390x, and you use

    debi --upgrade dbus_x.y-z_source+amd64+i386+s390x+all.changes

with i386 enabled as a foreign architecture and libdbus-1-3:amd64 and
libdbus-1-3:i386 both installed, it seems better to install both than
to fail to configure some packages.  However, installing dbus:i386 is
undesirable, assuming dbus:amd64 is the one currently installed.

The logic used when dpkg -Oi decides whether a package is already
installed appears to be: if it is Multi-Arch: same, see whether this
(package, architecture) pair is already installed; if not, see whether
this package is already installed for any architecture.

However, we don't want to try to install packages for unconfigured
foreign architectures (s390x in the example above).

Signed-off-by: Simon McVittie <smcv@debian.org>
Closes: #915668