mergechanges: Add a basic regression test
The valid .changes files used here are real data from my recent
xdg-desktop-portal upload, as produced with dpkg (>= 1.19.3)
and sbuild. In particular, the source changes file doesn't have Binary
or Description fields.

I've left the signature on the source changes file, so that it can be
used to verify that merging a signed changes file results in unsigned
output.

Signed-off-by: Simon McVittie <smcv@debian.org>