Devscripts::MkOrigtargz: Fix OpenPGP signature ASCII enarmor normalization
Properly fix the trailing armor line, not just the heading one (as it
was applying only to the first matched instance in the multi-line
string), and remove the Comment field from the pseudo-header block
(this is not sed and the «d» modifier to m// does not mean delete).

Switch to slurping the output into an array and handling each line
separately.

Signed-off-by: Guillem Jover <guillem@debian.org>