uscan: Don't recurse into directories named .git
dgit caches an unpacked copy of the package in .git/dgit/unpack/work.
When running uscan in a git checkout where dgit has been used, we don't
want to try to update using the (likely outdated) configuration in
.git/dgit/unpack/work/debian/watch, only the current configuration in
 ./debian/watch.

We probably also don't want to treat .git/refs/heads/ and .git/refs/tags/
as potential Debian packages, just because they happen to contain
a debian directory (when using DEP-14, and when using DEP-14 or gbp
defaults, respectively).

Signed-off-by: Simon McVittie <smcv@debian.org>