debi: Don't use apt install ./*.changes for --with-depends --upgrade
The semantics of --with-depends --upgrade have historically been that
we will upgrade existing packages mentioned in the given changes file,
and we will install their dependencies if necessary, but we will not
install new packages mentioned in the changes file.

For example, suppose we have installed the development tool piuparts,
but not the special-purpose package piuparts-master (which is only
intended to be installed on piuparts master servers). If we are
using debi to test a new version of piuparts that adds a dependency
on uidmap, then debi --upgrade --with-depends ../piuparts.changes
should install the new piuparts, and install its dependency uidmap,
but should not install piuparts-master.

apt --only-upgrade is not suitable here because it would prevent us
from installing uidmap; apt without --only-upgrade is also not suitable
because it would install piuparts-master. Fall back to using debpkg
instead.

Signed-off-by: Simon McVittie <smcv@debian.org>
Closes: #932963