debdiff: Optionally apply 3.0 (quilt) patches before comparison
This is particularly useful when comparing a Debian revision with
post-release fixes as patches with a new upstream version in which
those fixes have been merged upstream, or when comparing packages
that have moved from 1.0 to 3.0 (quilt) source format.

As well as implementing what I requested in #688219, this partially
addresses #939602 (but the submitter of #939602 additionally wanted
the default to change, which I have not done here).

In addition to the changes being applied to the upstream source code,
the comparison also (somewhat redundantly) includes the changes to
debian/patches. This allows a reviewer to see how any new changes break up
into individual 3.0 (quilt) patches, and an in particular lets a reviewer
see the (hopefully DEP-3) patch descriptions. If this is undesired,
the actual patches can be filtered out with `--exclude='*.patch'` or by
postprocessing with `filterdiff -x`.

For those familiar with `gbp-pq(1)`, this is approximately the same
comparison as the one obtained by importing the old and new patch
series with `gbp pq import`, and then comparing `patch-queue/old` with
`patch-queue/new`.

Signed-off-by: Simon McVittie <smcv@debian.org>
Closes: #688219