scripts/hardening-check.pl: Better CF protection detection
The current method of looking for particular instructions in the
disassembly is possibly misleading as the inclusion of these is not
necessarily indicative of actual CF protection as not all modules may have
been compiled with CET, in which case CET will not be enabled at
runtime. Instead, look for the a specific property in the notes section of
the ELF binary since this is added by the linker when all modules have CF
protection (ie CET).