mk-origtargz: Honor Files-ExcludeIgnored stanza
A common use-case for this can be seen in golang packages where one wants to
exclude most but not all subdirectories under the "vendor/" directory. Many
upstream packages aggressivly vendor dependencies, most of which are already
included in debian. In some cases, it is not worth the trouble to package them
seperately, in which case the new Files-ExcludeIgnored stanza can be very
convenient to leave those sources in.

Closes: #829646