grep-excuses --autopkgtests: Do not report OLD_PASS
I'm not sure precisely what this state means, but it does not seem to
mean a regression.  I got a mail from grep-excuses this morning
telling me that a test from dgit's autopkgtests, run for migration of
fakeroot, was OLD_PASS.  Now when I look I see the test has passed in
the meantime.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>