test_uscan_mangle: Use MGZREPACK to specify the expected file extension
This patch removes the GZREPACK variable to simplify the test sequence.

The patch also adds a minor workaround for testWatch4WebNonNativeMUT0(),
which only makes use of component tarballs, although it does download a
main upstream tarball.

The test sequence expects to find a main tarball with a .xz extension, but
because the main tarball is not repacked, the expected file is not found.
To work around this, we set STARBALL to the original tarball name.

No other test requires this special handling.